+++ src/PJ_cart.c
@@ -41,7 +41,7 @@
  *****************************************************************************/
 
 #define PJ_LIB__
-#include <proj.h>
+#include <proj_internal.h>
 #include <projects.h>
 #include <assert.h>
 #include <stddef.h>
