+++ src/PJ_horner.c
@@ -1,5 +1,5 @@
 #define PJ_LIB__
-#include <proj.h>
+#include <proj_internal.h>
 #include <projects.h>
 #include <assert.h>
 #include <stddef.h>
