+++ src/PJ_pipeline.c
@@ -96,7 +96,7 @@ Thomas Knudsen, thokn@sdfe.dk, 2016-05-20
 ********************************************************************************/
 
 #define PJ_LIB__
-#include <proj.h>
+#include <proj_internal.h>
 #include <projects.h>
 
 #include <assert.h>
